+++ src/FileMailer.php
@@ -39,21 +39,21 @@ final class FileMailer extends Mailer
     private $filenameCallback;
 
     /**
-     * @param MessageFactoryInterface $messageFactory The message factory instance.
      * @param MessageBodyRenderer $messageBodyRenderer The message body renderer instance.
-     * @param EventDispatcherInterface $eventDispatcher The event dispatcher instance.
      * @param string $path The path where message files located.
      * @param callable|null $filenameCallback A PHP callback that return a file name which will be used to save
      * the email message.
+     * @param MessageSettings|null $messageSettings The default and extra message settings.
+     * @param EventDispatcherInterface|null $eventDispatcher The event dispatcher instance.
      */
     public function __construct(
-        MessageFactoryInterface $messageFactory,
         MessageBodyRenderer $messageBodyRenderer,
         private string $path,
         callable $filenameCallback = null,
+        ?MessageSettings $messageSettings = null,
         ?EventDispatcherInterface $eventDispatcher = null,
     ) {
-        parent::__construct($messageFactory, $messageBodyRenderer, $eventDispatcher);
+        parent::__construct($messageBodyRenderer, $messageSettings, $eventDispatcher);
         $this->filenameCallback = $filenameCallback;
     }