+++ tests/MessageTest.php
@@ -17,22 +17,22 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
     public function testDefaultValues(): void
     {
         $message = new Message();
-        $this->assertSame('utf-8', $message->getCharset());
-        $this->assertSame([], $message->getFrom());
-        $this->assertSame([], $message->getTo());
-        $this->assertSame([], $message->getReplyTo());
-        $this->assertSame([], $message->getCc());
-        $this->assertSame([], $message->getBcc());
-        $this->assertSame('', $message->getSubject());
+        $this->assertNull($message->getCharset());
+        $this->assertNull($message->getFrom());
+        $this->assertNull($message->getTo());
+        $this->assertNull($message->getReplyTo());
+        $this->assertNull($message->getCc());
+        $this->assertNull($message->getBcc());
+        $this->assertNull($message->getSubject());
         $this->assertNull($message->getDate());
-        $this->assertSame(Priority::NORMAL, $message->getPriority());
-        $this->assertSame('', $message->getReturnPath());
-        $this->assertSame('', $message->getSender());
-        $this->assertSame('', $message->getTextBody());
-        $this->assertSame('', $message->getHtmlBody());
-        $this->assertSame([], $message->getAttachments());
-        $this->assertSame([], $message->getEmbeddings());
-        $this->assertSame([], $message->getHeaders());
+        $this->assertNull($message->getPriority());
+        $this->assertNull($message->getReturnPath());
+        $this->assertNull($message->getSender());
+        $this->assertNull($message->getTextBody());
+        $this->assertNull($message->getHtmlBody());
+        $this->assertNull($message->getAttachments());
+        $this->assertNull($message->getEmbeddings());
+        $this->assertNull($message->getHeaders());
     }
 
     public function testWithCharset(): void
@@ -47,10 +47,40 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
         $this->assertSame('test@example.com', $message->getFrom());
     }
 
+    public function testWithAddedFrom(): void
+    {
+        $sourceMessage = new Message(from: 'test@example.com');
+
+        $message = $sourceMessage->withAddedFrom('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(['test@example.com', 'mark@example.com'], $message->getFrom());
+    }
+
+    public function testWithAddedFromToMessageWithoutFrom(): void
+    {
+        $sourceMessage = new Message();
+
+        $message = $sourceMessage->withAddedFrom('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame('mark@example.com', $message->getFrom());
+    }
+
     public function testWithTo(): void
     {
         $message = (new Message())->withTo('test@example.com');
         $this->assertSame('test@example.com', $message->getTo());
+    }
+
+    public function testWithAddedTo(): void
+    {
+        $sourceMessage = new Message(to: 'test@example.com');
+
+        $message = $sourceMessage->withAddedTo('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(['test@example.com', 'mark@example.com'], $message->getTo());
     }
 
     public function testWithReplyTo(): void
@@ -59,16 +89,46 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
         $this->assertSame('test@example.com', $message->getReplyTo());
     }
 
+    public function testWithAddedReplyTo(): void
+    {
+        $sourceMessage = new Message(replyTo: 'test@example.com');
+
+        $message = $sourceMessage->withAddedReplyTo('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(['test@example.com', 'mark@example.com'], $message->getReplyTo());
+    }
+
     public function testWithCc(): void
     {
         $message = (new Message())->withCc('test@example.com');
         $this->assertSame('test@example.com', $message->getCc());
     }
 
+    public function testWithAddedCc(): void
+    {
+        $sourceMessage = new Message(cc: 'test@example.com');
+
+        $message = $sourceMessage->withAddedCc('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(['test@example.com', 'mark@example.com'], $message->getCc());
+    }
+
     public function testWithBcc(): void
     {
         $message = (new Message())->withBcc('test@example.com');
         $this->assertSame('test@example.com', $message->getBcc());
+    }
+
+    public function testWithAddedBcc(): void
+    {
+        $sourceMessage = new Message(bcc: 'test@example.com');
+
+        $message = $sourceMessage->withAddedBcc('mark@example.com');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(['test@example.com', 'mark@example.com'], $message->getBcc());
     }
 
     public function testWithSubject(): void
@@ -82,13 +142,18 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
         return [
             'immutable' => [new DateTimeImmutable()],
             'mutable' => [new DateTime()],
+            'null' => [null],
         ];
     }
 
     #[DataProvider('dataDate')]
-    public function testWithDate(DateTimeInterface $date): void
+    public function testWithDate(?DateTimeInterface $date): void
     {
-        $message = (new Message())->withDate($date);
+        $sourceMessage = new Message(date: new DateTimeImmutable());
+
+        $message = $sourceMessage->withDate($date);
+
+        $this->assertNotSame($message, $sourceMessage);
         $this->assertEquals($date, $message->getDate());
     }
 
@@ -102,6 +167,26 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
 
         $this->assertSame([], $message->getHeader('X-Not-Exists'));
         $this->assertSame(['a', 'b'], $message->getHeader('X-Test'));
+    }
+
+    public function testWithoutAttachments(): void
+    {
+        $sourceMessage = new Message(attachments: [File::fromContent('hello')]);
+
+        $message = $sourceMessage->withoutAttachments();
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertNull($message->getAttachments());
+    }
+
+    public function testWithoutEmbeddings(): void
+    {
+        $sourceMessage = new Message(embeddings: [File::fromContent('hello')]);
+
+        $message = $sourceMessage->withoutEmbeddings();
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertNull($message->getEmbeddings());
     }
 
     public function testToString(): void
@@ -123,6 +208,96 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
         );
     }
 
+    public function testWithHeaders(): void
+    {
+        $sourceMessage = new Message(headers: ['X-Test' => '1']);
+
+        $message = $sourceMessage->withHeaders([
+            'X-Origin' => ['0', '1'],
+            'X-Pass' => 'pass',
+        ]);
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame(
+            [
+                'X-Origin' => ['0', '1'],
+                'X-Pass' => ['pass'],
+            ],
+            $message->getHeaders(),
+        );
+    }
+
+    public function testWithHtmlBody(): void
+    {
+        $sourceMessage = new Message(htmlBody: '<b>Hello</b>');
+
+        $message = $sourceMessage->withHtmlBody('<i>Test</i>');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame('<i>Test</i>', $message->getHtmlBody());
+    }
+
+    public function testWithTextBody(): void
+    {
+        $sourceMessage = new Message(textBody: 'Hello');
+
+        $message = $sourceMessage->withTextBody('Test');
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame('Test', $message->getTextBody());
+    }
+
+    public function testWithAddedAttachments(): void
+    {
+        $file1 = File::fromContent('1');
+        $file2 = File::fromContent('2');
+        $file3 = File::fromContent('3');
+        $sourceMessage = new Message(attachments: [$file1]);
+
+        $message = $sourceMessage->withAddedAttachments($file2, $file3);
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame([$file1, $file2, $file3], $message->getAttachments());
+    }
+
+    public function testWithAddedEmbeddings(): void
+    {
+        $file1 = File::fromContent('1');
+        $file2 = File::fromContent('2');
+        $file3 = File::fromContent('3');
+        $sourceMessage = new Message(embeddings: [$file1]);
+
+        $message = $sourceMessage->withAddedEmbeddings($file2, $file3);
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame([$file1, $file2, $file3], $message->getEmbeddings());
+    }
+
+    public static function dataAddedHeader(): iterable
+    {
+        yield [
+            ['X-Test' => ['0', '1']],
+            'X-Test',
+            '1',
+        ];
+        yield [
+            ['X-Test' => ['0'], 'X-Origin' => ['on']],
+            'X-Origin',
+            'on',
+        ];
+    }
+
+    #[DataProvider('dataAddedHeader')]
+    public function testWithAddedHeader(array $expected, string $headerName, string $headerValue): void
+    {
+        $sourceMessage = new Message(headers: ['X-Test' => '0']);
+
+        $message = $sourceMessage->withAddedHeader($headerName, $headerValue);
+
+        $this->assertNotSame($message, $sourceMessage);
+        $this->assertSame($expected, $message->getHeaders());
+    }
+
     public function testImmutability(): void
     {
         $message = new Message();
@@ -138,11 +313,7 @@ final class MessageTest extends \PHPUnit\Framework\TestCase
         $this->assertNotSame($message, $message->withSender(''));
         $this->assertNotSame($message, $message->withPriority(Priority::HIGHEST));
         $this->assertNotSame($message, $message->withAttachments(File::fromContent('')));
-        $this->assertNotSame($message, $message->withAddedAttachments());
         $this->assertNotSame($message, $message->withEmbeddings(File::fromContent('')));
-        $this->assertNotSame($message, $message->withAddedEmbeddings());
-        $this->assertNotSame($message, $message->withAddedHeader('X-Test', '0'));
         $this->assertNotSame($message, $message->withHeader('X-Test', '0'));
-        $this->assertNotSame($message, $message->withHeaders([]));
     }
 }