+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Chg #96: Change order of constructor parameters in `Mailer` and `FailMailer` (@vjik)
 - Chg #98: Make `MessageBodyTemplate` simple DTO with public readonly properties (@vjik)
 - Enh #99: Add `yiisoft/view` of version `^11` support (@vjik)
+- Chg #100: Remove `MessageFactoryInterface` and `MessageFactory`, add `MessageSettings` instead it (@vjik)
 
 ## 5.1.0 July 02, 2024